+++ yallib/models/user.py
@@ -1,43 +1,8 @@
-from django.db import models
-from datetime import datetime
 from django.contrib.auth.models import (
     BaseUserManager, AbstractBaseUser
     )
-class BookManager(models.Manager):
-    def get_by_natural_key(self, a, b, c):
-        return self.get(first_name=a,last_name=b,date_birth=c)
-
-class BaseModel(models.Model):
-    created = models.DateTimeField(default=datetime.now, blank=True)
-    changed = models.DateTimeField(auto_now=True)
-    class Meta:
-        abstract = True
-
-class Author(BaseModel):
-    objects = BookManager()
-    first_name = models.CharField(max_length=100)
-    last_name = models.CharField(max_length=100)
-    date_birth = models.DateField()
-
-    class Meta:
-        unique_together = (("first_name", "last_name", "date_birth"))
-
-class Publication(BaseModel):
-    caption = models.CharField(max_length=100)
-    authors = models.ManyToManyField(Author,related_name="authors")
-    class Meta:
-        abstract = True
-
-class Genre(BaseModel):
-    name = models.CharField(max_length=100)
-
-class Book(Publication):
-    genres = models.ManyToManyField(Author,related_name="genres")
-###############################################################################
-
- # https://docs.djangoproject.com/en/1.7/topics/auth/customizing/#a-full-example
 class UserManager(BaseUserManager):
-    def create_user(self,email,password=None):
+    def create_user(self, email, password=None):
         """
         Создаем и сохраняем Юзера с полученным емаил и парол
         """
@@ -74,7 +39,7 @@ class UserManager(BaseUserManager):
         return user
 class User(AbstractBaseUser): # создаем таблицу пользователь в БД
 
-    email = models.EmailField(verbose_name = 'email address',max_length=100,unique = True)
+    email = models.EmailField(verbose_name = 'email address', max_length=100, unique = True)
     active = models.BooleanField(default=True)
     staff = models.BooleanField(default=False) # admin User
     admin = models.BooleanField(default=False) # super User
